# store: Added benchmarks for .Series and expanded matchers.