@bwplotka 2020-01-29
You can ignore CI error, it's because I need to rebase on https://github.com/thanos-io/thanos/pull/2077 once done.