@bwplotka 2020-01-30
First iteration of `.Series` result is fun! (: 

```
bash _dev/bench_series.sh 0
Running BenchmarkSeries/binary_header run 0-series
goos: linux
goarch: amd64
pkg: github.com/thanos-io/thanos/pkg/store
BenchmarkSeries_MillionsSeriesWithOneSample/binary_header/1of10000000-12         	      10	3170961965 ns/op	1577961656 B/op	20076938 allocs/op
BenchmarkSeries_MillionsSeriesWithOneSample/binary_header/10of10000000-12        	      10	3233990066 ns/op	1577881177 B/op	20076844 allocs/op
BenchmarkSeries_MillionsSeriesWithOneSample/binary_header/100of10000000-12       	      10	3179363645 ns/op	1577940600 B/op	20077338 allocs/op
BenchmarkSeries_MillionsSeriesWithOneSample/binary_header/1000of10000000-12      	      10	3118202426 ns/op	1578423198 B/op	20081876 allocs/op
BenchmarkSeries_MillionsSeriesWithOneSample/binary_header/10000of10000000-12     	      10	3139396100 ns/op	1587140179 B/op	20127240 allocs/op
BenchmarkSeries_MillionsSeriesWithOneSample/binary_header/100000of10000000-12    	      10	3253384227 ns/op	1671065876 B/op	20580979 allocs/op
BenchmarkSeries_MillionsSeriesWithOneSample/binary_header/1000000of10000000-12   	       8	4105534012 ns/op	2518368762 B/op	25115331 allocs/op
bucket_test.go:1223: unexpected error: rpc error: code = Aborted desc = fetch series for block 01DZV3V8TSFV4008VD6D6YVNKP: preload chunks: read range for 0: allocate chunk bytes: pool exhausted

--- FAIL: BenchmarkSeries_MillionsSeriesWithOneSample/binary_header/10000000of10000000
--- FAIL: BenchmarkSeries_MillionsSeriesWithOneSample/binary_header
--- FAIL: BenchmarkSeries_MillionsSeriesWithOneSample
FAIL
exit status 1
FAIL	github.com/thanos-io/thanos/pkg/store	390.390s
FAIL
Running BenchmarkSeries/index_json run 0-series
goos: linux
goarch: amd64
pkg: github.com/thanos-io/thanos/pkg/store
BenchmarkSeries_MillionsSeriesWithOneSample/index_json/1of10000000-12         	      15	2356337341 ns/op	1437954038 B/op	15076911 allocs/op
BenchmarkSeries_MillionsSeriesWithOneSample/index_json/10of10000000-12        	      15	2371328532 ns/op	1437874976 B/op	15076814 allocs/op
BenchmarkSeries_MillionsSeriesWithOneSample/index_json/100of10000000-12       	      15	2387921632 ns/op	1437963037 B/op	15077418 allocs/op
BenchmarkSeries_MillionsSeriesWithOneSample/index_json/1000of10000000-12      	      15	2362966154 ns/op	1438438411 B/op	15081966 allocs/op
BenchmarkSeries_MillionsSeriesWithOneSample/index_json/10000of10000000-12     	      15	2395617626 ns/op	1447102662 B/op	15127229 allocs/op
BenchmarkSeries_MillionsSeriesWithOneSample/index_json/100000of10000000-12    	      15	2404596601 ns/op	1530925085 B/op	15580972 allocs/op
BenchmarkSeries_MillionsSeriesWithOneSample/index_json/1000000of10000000-12   	      12	3237704906 ns/op	2374237739 B/op	20115173 allocs/op
bucket_test.go:1223: unexpected error: rpc error: code = Aborted desc = fetch series for block 01DZV49S7ZXV2YF4VS4NQ1DSZS: preload chunks: read range for 0: allocate chunk bytes: pool exhausted
```
2GB of memory to fetch 1mln series and that is WITH the chunk bytes pool (: Trying without...