@bwplotka 2020-01-30
Without pooling bytes for chunks it' madness:

```
Running BenchmarkSeries/binary_header run 1-series
goos: linux
goarch: amd64
pkg: github.com/thanos-io/thanos/pkg/store
BenchmarkSeries/binary_header/1of10000000-12         	      10	3284197505 ns/op	1577936804 B/op	20076854 allocs/op
BenchmarkSeries/binary_header/10of10000000-12        	      10	3295759539 ns/op	1577945545 B/op	20076895 allocs/op
BenchmarkSeries/binary_header/100of10000000-12       	      10	3302932189 ns/op	1578028665 B/op	20077475 allocs/op
BenchmarkSeries/binary_header/1000of10000000-12      	      10	3275058090 ns/op	1578476056 B/op	20081734 allocs/op
BenchmarkSeries/binary_header/10000of10000000-12     	      10	3254601925 ns/op	1587646488 B/op	20127175 allocs/op
BenchmarkSeries/binary_header/100000of10000000-12    	      10	3366360233 ns/op	1676253709 B/op	20580883 allocs/op
BenchmarkSeries/binary_header/1000000of10000000-12   	       8	4232465798 ns/op	2566112362 B/op	25115298 allocs/op
BenchmarkSeries/binary_header/10000000of10000000-12  	       3	13466568311 ns/op	15881951480 B/op	130613948 allocs/op
PASS
ok  	github.com/thanos-io/thanos/pkg/store	480.663s
Running BenchmarkSeries/index_json run 1-series
goos: linux
goarch: amd64
pkg: github.com/thanos-io/thanos/pkg/store
BenchmarkSeries/index_json/1of10000000-12         	      15	2424057657 ns/op	1437937921 B/op	15076856 allocs/op
BenchmarkSeries/index_json/10of10000000-12        	      13	2538203730 ns/op	1437956158 B/op	15076931 allocs/op
BenchmarkSeries/index_json/100of10000000-12       	      13	2447718936 ns/op	1437993884 B/op	15077353 allocs/op
BenchmarkSeries/index_json/1000of10000000-12      	      15	2507760852 ns/op	1438558841 B/op	15082019 allocs/op
BenchmarkSeries/index_json/10000of10000000-12     	      13	2488399436 ns/op	1447666612 B/op	15127243 allocs/op
BenchmarkSeries/index_json/100000of10000000-12    	      14	2664945098 ns/op	1536270808 B/op	15580945 allocs/op
BenchmarkSeries/index_json/1000000of10000000-12   	      10	3407783584 ns/op	2426086508 B/op	20115231 allocs/op
BenchmarkSeries/index_json/10000000of10000000-12  	signal: killed
FAIL	github.com/thanos-io/thanos/pkg/store	511.756s
FAIL
```
It looks good: index-cache did not complete due to OOM (: vs the index-header manage to do it 3 times for 30s :heart: