@bwplotka 2020-01-30
Flame graph for tests with chunk bytes pool (so default for Thanos):

![image](https://user-images.githubusercontent.com/6950331/73458225-d9555a80-436c-11ea-929b-9f6a6144699a.png)

Interesting questions / major mem usages:
* read bytes from object storage takes a lot of space: I think we need to reuse more bytes / pool - this will help a lot :+1: 
* It would be nice to cache / pool symbols at some point... They take significant amount as well.
* we are sorting.. I guess labels here:

```
sort.Slice(s.lset, func(i, j int) bool {
			return s.lset[i].Name < s.lset[j].Name
})
```
This is becasue we are adding external labels. It takes as well some mem, but really not good idea how to avoid this - I think labels has to be sorted overall.