@bwplotka 2020-01-30
OK, this is ready go. In separate PR I will remove index_json cases. We want to get rid of this, so no point spending time in benchmarking this. PTAL @GiedriusS @brancz @squat @metalmatze @kakkoyun